refactor(copilot-app): extract workflow integrator + fold Copilot review nits
The copilot-app target deploys *.prompt.md as Copilot App workflow rows
(SQLite + optional WS-IPC) instead of files. That path shared NOTHING
with the file-based prompt deploy except the source artefact, but it
was squatting inside PromptIntegrator (~300 lines of unrelated logic
plus its own module-level frontmatter helpers). Extract into a sibling
integrator so the file matches how agent_integrator / command_integrator
/ hook_integrator / instruction_integrator are organised.

Refactor:

* New src/apm_cli/integration/copilot_app_workflow_integrator.py with
  CopilotAppWorkflowIntegrator(BaseIntegrator). Inherits BaseIntegrator
  for find_files_by_glob only -- the file-based collision /
  link-resolution machinery is irrelevant on the workflow surface
  (deploy_workflow is an UPSERT keyed on a namespaced id; sync deletes
  by id from the lockfile).
* PromptIntegrator.integrate_prompts_for_target and .sync_for_target
  keep one trivially small branch per method that delegates to the
  workflow integrator -- grep copilot-app in prompt_integrator.py and
  one call site appears for each direction.
* Module-level frontmatter helpers (_is_workflow_shape, Schedule,
  _parse_workflow_frontmatter, _parse_schedule, _derive_package_owner)
  move with the path that uses them. Re-exported from prompt_integrator
  for back-compat with tests / external callers.
* copilot_app_db / copilot_app_project / copilot_app_ws stay put --
  they are already at the right level. Only prompt_integrator was
  oversized.
* Test fixtures unchanged; the integration test that patches
  copilot_app_ws.WsClient.create_project_from_path still works because
  the import path did not move.

Folded Copilot reviewer nits (PR #1431):

* B1: when derive_repo_context(project_root) returns None AND workflow
  prompts are present AND scope is not user, emit a parallel warn to
  the existing --global warn. project_id=NULL workflows have the same
  CWD=~/.copilot pivot risk; the user deserves the same heads-up plus
  the fix recipe (run inside a git repo or attach from the App UI).
* B2: WsClient.create_project_from_path docstring rewritten to match
  the implementation -- was_created is inferred from the reply type
  (project_created -> True, project_updated -> False) via
  _extract_project_fields, not 'always True'. Restart hint only fires
  on the project_created branch.
* B3: _find_repo_root docstring no longer claims symlinks-in-the-chain
  are blocked. The cheap fix: drop the false claim. The 'right' fix
  (walk + reject symlinked parents) would break macOS /tmp ->
  /private/tmp and other legitimate symlinked-ancestor setups, and the
  threat model (user's own CWD, not adversary input) does not justify
  the cost. The .git-marker is_symlink() check stays as
  defence-in-depth.

Behavior change: one new warn diagnostic when copilot-app workflows
install without a detected git repo. Three pre-existing unit tests in
test_copilot_app_error_ux.py updated to filter the new warn out of
their per-prompt deploy-error counts (the no-repo warn fires once for
the whole install; deploy-error warns are per-prompt).

Test gate: 54 targeted + 2766 broad -- all pass.
