+++ src/apm_cli/integration/prompt_integrator.py
@@ -86,12 +86,19 @@ class PromptIntegrator(BaseIntegrator):
         # lifecycle for this target -- it does not share the file-based
         # collision / link-resolution machinery.
         if target.name == "copilot-app":
+            # Delegate to the workflow integrator -- file-based and
+            # SQLite-row deploy share NOTHING except the source
+            # artefact. See copilot_app_workflow_integrator.py.
+            from apm_cli.integration.copilot_app_workflow_integrator import (
+                CopilotAppWorkflowIntegrator,
+            )
+
             # Detect --global / user-scope by name; we accept either the
             # InstallScope enum or a string. Avoids a hard import cycle.
             user_scope = False
             if scope is not None:
                 user_scope = getattr(scope, "name", str(scope)).upper() == "USER"
-            return self._integrate_prompts_for_copilot_app(
+            return CopilotAppWorkflowIntegrator().integrate(
                 target,
                 package_info,
                 project_root=project_root,
@@ -124,7 +131,11 @@ class PromptIntegrator(BaseIntegrator):
             return {"files_removed": 0, "errors": 0}
 
         if target.name == "copilot-app":
-            return self._sync_copilot_app(managed_files or set())
+            from apm_cli.integration.copilot_app_workflow_integrator import (
+                CopilotAppWorkflowIntegrator,
+            )
+
+            return CopilotAppWorkflowIntegrator().sync(managed_files or set())
 
         effective_root = mapping.deploy_root or target.root_dir
         prefix = f"{effective_root}/{mapping.subdir}/"
@@ -139,7 +150,10 @@ class PromptIntegrator(BaseIntegrator):
         )
 
     # ------------------------------------------------------------------
-    # copilot-app SQLite path
+    # copilot-app SQLite path -- MOVED to copilot_app_workflow_integrator.py.
+    # The methods below are kept as thin shims for back-compat with any
+    # external caller / test that imported them directly. They will be
+    # removed in a future major.
     # ------------------------------------------------------------------
 
     def _integrate_prompts_for_copilot_app(
@@ -152,288 +166,34 @@ class PromptIntegrator(BaseIntegrator):
         force: bool,
         diagnostics,
     ) -> IntegrationResult:
-        """Deploy workflow-shape prompts as Copilot App workflow rows.
+        """Back-compat shim -- forwards to CopilotAppWorkflowIntegrator.integrate.
 
-        Hybrid dispatch:
-
-        1. Detect the repo at ``project_root``. Derive a stable
-           ``RepoContext`` (name, github owner/repo, default branch) so
-           workflow rows scope to a real ``projects`` row instead of
-           being orphaned at the root.
-        2. If ``user_scope`` is set AND any workflow-shape prompts are
-           present, emit a warn-and-proceed diagnostic: workflows run
-           with CWD=~/.copilot in global mode, which is almost never
-           what the user wants. We still deploy so global skills /
-           commands keep working; the user can attach the row to a
-           project from the App UI.
-        3. Resolve the ``project_id`` ONCE: if the App is running
-           (``ws_available``), try ``WsClient.create_project_from_path``
-           so the project row goes through the App's own validation
-           (owner/repo detection, default branch, account binding) and
-           is already known to the webview. On any ``WsError`` (or if
-           the App is closed) fall through to direct SQLite via
-           ``resolve_or_register_project_sqlite``.
-        4. Write every workflow row via direct SQLite
-           (``copilot_app_db.deploy_workflow``), stamped with the
-           resolved ``project_id`` and a namespaced
-           ``owner/pkg/stem`` id so the lockfile stays stable across
-           runs and across the WS-vs-SQLite project-resolution branch.
-        5. If the project was freshly created (either branch), emit
-           the one-time restart hint -- the webview's projectStore
-           does not currently refresh on a new ``projects`` row.
-
-        Workflow-shape (per ``_is_workflow_shape``) prompts deploy
-        here.  Plain-shape prompts (no execution-affecting frontmatter
-        keys) are a hard error at this target: the user explicitly
-        opted into ``copilot-app`` for this package, and a plain
-        prompt cannot possibly be a workflow.  Surfacing the mismatch
-        loudly beats silently skipping the file and leaving the user
-        wondering why nothing landed in the App.
-
-        The DB module enforces ``enabled = 0`` on insert; any frontmatter
-        ``enabled`` field, if present, is ignored.  This is a hard
-        contract: third-party packages cannot auto-run anything on the
-        user's machine.
+        The real implementation moved to
+        apm_cli.integration.copilot_app_workflow_integrator. Kept here so
+        external callers / tests that still reference the method-on-PromptIntegrator
+        attribute continue to work. New code should construct
+        CopilotAppWorkflowIntegrator directly.
         """
-        import frontmatter
-
-        from apm_cli.integration.copilot_app_db import (
-            CopilotAppDbError,
-            WorkflowRow,
-            deploy_workflow,
-            namespaced_id,
-            resolve_copilot_app_db_path,
-        )
-        from apm_cli.integration.copilot_app_project import (
-            derive_repo_context,
-            resolve_or_register_project_sqlite,
-        )
-        from apm_cli.integration.copilot_app_ws import (
-            WsAppNotRunning,
-            WsAuthError,
-            WsClient,
-            WsError,
-            ws_available,
+        from apm_cli.integration.copilot_app_workflow_integrator import (
+            CopilotAppWorkflowIntegrator,
         )
 
-        db_path = resolve_copilot_app_db_path()
-        if db_path is None:
-            return IntegrationResult(0, 0, 0, [])
-
-        owner = _derive_package_owner(package_info)
-        pkg_name = package_info.package.name
-
-        # ------------------------------------------------------------------
-        # Parse all candidate prompts up front so we can decide on the
-        # global-scope warning with one pass.
-        # ------------------------------------------------------------------
-        parsed: list[tuple[Path, object, object]] = []
-        files_skipped = 0
-        for source_file in self.find_prompt_files(package_info.install_path):
-            if source_file.is_symlink():
-                if diagnostics is not None:
-                    diagnostics.warn(
-                        message=f"Refusing to read symlink prompt: {source_file}",
-                        package=pkg_name,
-                    )
-                files_skipped += 1
-                continue
-            post = frontmatter.load(str(source_file))
-            if not _is_workflow_shape(post.metadata):
-                if diagnostics is not None:
-                    diagnostics.warn(
-                        message=(
-                            f"Copilot App: {source_file.name} has no workflow frontmatter "
-                            "(missing one of: interval, schedule_hour, schedule_day). "
-                            "Add `interval: manual` to deploy it as a manual-trigger "
-                            "workflow, or unset --target copilot-app."
-                        ),
-                        package=pkg_name,
-                    )
-                files_skipped += 1
-                continue
-            try:
-                schedule = _parse_workflow_frontmatter(post.metadata)
-            except ValueError as exc:
-                if diagnostics is not None:
-                    diagnostics.warn(
-                        message=f"Invalid workflow frontmatter in {source_file.name}: {exc}",
-                        package=pkg_name,
-                    )
-                files_skipped += 1
-                continue
-            parsed.append((source_file, post, schedule))
-
-        # ------------------------------------------------------------------
-        # --global + workflow-shape: warn but proceed.
-        # ------------------------------------------------------------------
-        if user_scope and parsed and diagnostics is not None:
-            diagnostics.warn(
-                message=(
-                    "Copilot App workflows installed with --global run with CWD=~/.copilot, "
-                    "not a project. Attach the workflow to a project from the App's Workflows "
-                    "tab to fix this, or re-run `apm install` from a repo without --global. "
-                    "See https://aka.ms/apm/copilot-app-global for details."
-                ),
-                package=pkg_name,
-            )
-
-        # ------------------------------------------------------------------
-        # Resolve project_id ONCE -- WS (preferred) then SQLite fallback.
-        # ------------------------------------------------------------------
-        repo_ctx = derive_repo_context(project_root)
-        repo_suffix = f" ({repo_ctx.repo_name})" if repo_ctx is not None else ""
-        project_id: str | None = None
-        was_created = False
-
-        if repo_ctx is not None:
-            if ws_available():
-                try:
-                    with WsClient() as client:
-                        project = client.create_project_from_path(repo_ctx.repo_root)
-                    project_id = project.project_id
-                    was_created = project.was_created
-                except WsAppNotRunning:
-                    # Race: ws_available() saw the port files but the
-                    # App closed between probe and handshake. Silent
-                    # fallback -- closed App is the normal off-state.
-                    pass
-                except WsAuthError:
-                    # Stale ``ws.token`` -- most often because the App
-                    # rotated its token at restart. We fall back to
-                    # SQLite silently rather than nag on every install:
-                    # the user-visible signal is already the restart
-                    # hint emitted further down, and the SQLite path
-                    # produces an identical project row.
-                    pass
-                except WsError as exc:
-                    if diagnostics is not None:
-                        diagnostics.warn(
-                            message=(
-                                f"Could not reach the running Copilot App "
-                                f"({exc}). Registering the project directly "
-                                "in the App database instead."
-                            ),
-                            package=pkg_name,
-                        )
-
-            if project_id is None:
-                try:
-                    resolved = resolve_or_register_project_sqlite(db_path, repo_ctx)
-                    project_id = resolved.project_id
-                    was_created = resolved.was_created
-                except CopilotAppDbError as exc:
-                    if diagnostics is not None:
-                        diagnostics.warn(
-                            message=(
-                                f"Could not register project for Copilot App: {exc}. "
-                                "Workflows will be installed without a project binding."
-                            ),
-                            package=pkg_name,
-                        )
-
-        if was_created and diagnostics is not None:
-            # See github/github-app#5483 -- the App webview does not
-            # currently refresh on externally-inserted ``projects``
-            # rows. A one-time restart wires the new project into the
-            # UI; subsequent installs into the same repo are silent.
-            diagnostics.info(
-                message=(
-                    "Registered a new Copilot App project for this repo. "
-                    "Restart the Copilot App once so the new project appears in the UI "
-                    "(see github/github-app#5483)."
-                ),
-                package=pkg_name,
-            )
-
-        # ------------------------------------------------------------------
-        # Workflow rows -- always SQLite, regardless of project source.
-        # Keeps lockfile ids namespaced and stable.
-        # ------------------------------------------------------------------
-        synthetic_root = db_path.parent / "workflows"
-        files_integrated = 0
-        target_paths: list[Path] = []
-        for source_file, post, schedule in parsed:
-            prompt_stem = source_file.name.removesuffix(".prompt.md")
-            wf_id = namespaced_id(owner, pkg_name, prompt_stem)
-            base_name = post.metadata.get("name") or prompt_stem
-            display_name = f"{base_name}{repo_suffix}"
-            row = WorkflowRow(
-                id=wf_id,
-                name=str(display_name),
-                prompt=post.content,
-                interval=schedule.interval,
-                schedule_hour=schedule.schedule_hour,
-                schedule_day=schedule.schedule_day,
-                enabled=0,  # ALWAYS disabled on install -- contract.
-                model=schedule.model,
-                reasoning_effort=schedule.reasoning_effort,
-                mode=schedule.mode,
-                project_id=project_id,
-            )
-            try:
-                deploy_workflow(db_path, row)
-            except CopilotAppDbError as exc:
-                if diagnostics is not None:
-                    diagnostics.warn(
-                        message=f"Could not deploy {prompt_stem!r} to Copilot App: {exc}",
-                        package=pkg_name,
-                    )
-                files_skipped += 1
-                continue
-            files_integrated += 1
-            target_paths.append(synthetic_root / wf_id)
-
-        return IntegrationResult(
-            files_integrated=files_integrated,
-            files_updated=0,
-            files_skipped=files_skipped,
-            target_paths=target_paths,
-            links_resolved=0,
-            files_adopted=0,
+        return CopilotAppWorkflowIntegrator().integrate(
+            target,
+            package_info,
+            project_root=project_root,
+            user_scope=user_scope,
+            force=force,
+            diagnostics=diagnostics,
         )
 
     def _sync_copilot_app(self, managed_files: set[str]) -> dict[str, int]:
-        """Remove Copilot App workflow rows referenced by *managed_files*.
-
-        Filters the input set to ``copilot-app-db://workflows/`` URIs,
-        decodes the workflow ids, and deletes them in a single
-        transaction.  Non-APM-namespaced ids are rejected by
-        ``copilot_app_db.delete_workflows`` for defence in depth.
-        """
-        from apm_cli.integration.copilot_app_db import (
-            COPILOT_APP_LOCKFILE_PREFIX,
-            CopilotAppDbError,
-            delete_workflows,
-            from_lockfile_uri,
-            is_copilot_app_uri,
-            resolve_copilot_app_db_path,
+        """Back-compat shim -- forwards to CopilotAppWorkflowIntegrator.sync."""
+        from apm_cli.integration.copilot_app_workflow_integrator import (
+            CopilotAppWorkflowIntegrator,
         )
 
-        ids: list[str] = []
-        for entry in managed_files:
-            if not is_copilot_app_uri(entry):
-                continue
-            if not entry.startswith(COPILOT_APP_LOCKFILE_PREFIX):
-                continue
-            try:
-                ids.append(from_lockfile_uri(entry))
-            except ValueError:
-                # Malformed entry -- skip rather than fail uninstall.
-                continue
-        if not ids:
-            return {"files_removed": 0, "errors": 0}
-
-        db_path = resolve_copilot_app_db_path()
-        if db_path is None:
-            # DB gone -- nothing to remove; treat as success (idempotent).
-            return {"files_removed": 0, "errors": 0}
-
-        try:
-            removed = delete_workflows(db_path, ids)
-        except CopilotAppDbError:
-            return {"files_removed": 0, "errors": 1}
-        return {"files_removed": removed, "errors": 0}
+        return CopilotAppWorkflowIntegrator().sync(managed_files)
 
     # ------------------------------------------------------------------
     # Legacy per-target API (DEPRECATED)
@@ -581,159 +341,26 @@ class PromptIntegrator(BaseIntegrator):
 
 
 # ---------------------------------------------------------------------------
-# Schedule frontmatter helpers (copilot-app target)
+# Schedule frontmatter helpers -- MOVED to copilot_app_workflow_integrator.
+#
+# Re-exported here for back-compat with tests / external consumers that still
+# import them from this module. New code should import from
+# apm_cli.integration.copilot_app_workflow_integrator directly.
 # ---------------------------------------------------------------------------
 
-from dataclasses import dataclass  # noqa: E402
+from apm_cli.integration.copilot_app_workflow_integrator import (  # noqa: E402
+    Schedule,
+    _derive_package_owner,
+    _is_workflow_shape,
+    _parse_schedule,
+    _parse_workflow_frontmatter,
+)
 
-_VALID_SCHEDULE_INTERVALS: frozenset[str] = frozenset({"manual", "hourly", "daily", "weekly"})
-_VALID_SCHEDULE_MODES: frozenset[str] = frozenset({"interactive", "plan"})
-"""Mirror of ``copilot_app_db._VALID_MODES``.  ``autopilot`` is
-deliberately omitted -- see that module's docstring for the
-secure-by-default rationale."""
-
-# Top-level frontmatter keys that mark a ``.prompt.md`` as a "workflow"
-# (i.e. a prompt with execution metadata, destined for the Copilot App
-# DB rather than slash-command file targets).  Touching ANY of these
-# keys at the top level of the frontmatter flips the dispatch shape.
-#
-# This is the Option B "dispatch by shape" predicate -- one folder
-# (``.apm/prompts/``), one extension (``.prompt.md``), one integrator.
-# A file with these keys ships to ``copilot-app`` and is skipped by
-# slash-command targets; a file without them ships to slash-command
-# targets and hard-errors at ``copilot-app``.
-_WORKFLOW_SHAPE_KEYS: frozenset[str] = frozenset({"interval", "schedule_hour", "schedule_day"})
-
-
-def _is_workflow_shape(frontmatter_meta: dict) -> bool:
-    """Return True iff *frontmatter_meta* declares Copilot App execution metadata.
-
-    Used to decide which target(s) a ``.prompt.md`` file is destined
-    for.  The check is intentionally a SHAPE check rather than a flag
-    -- authors do not opt in with a sentinel; the presence of an
-    execution-affecting key is the opt-in.
-
-    Only ``interval``, ``schedule_hour``, ``schedule_day`` are
-    unambiguous workflow markers.  ``mode``, ``model``, and
-    ``reasoning_effort`` are deliberately EXCLUDED because they overload
-    with plain slash-command prompts: VSCode / Copilot prompts use
-    ``mode: agent|ask|edit``, can pin a ``model``, and can hint
-    ``reasoning_effort``.  Treating those keys as workflow markers
-    would mis-route ordinary slash commands to the App DB.  Authors who
-    want a manual-only workflow opt in with the explicit
-    ``interval: manual``.
-    """
-    if not isinstance(frontmatter_meta, dict):
-        return False
-    return any(k in frontmatter_meta for k in _WORKFLOW_SHAPE_KEYS)
-
-
-@dataclass(frozen=True)
-class Schedule:
-    """Validated representation of a prompt's workflow frontmatter.
-
-    All fields are pre-validated against the same constraints the
-    Copilot App's ``workflows`` schema enforces, so deploy time never
-    surfaces a raw SQLite ``CHECK`` violation to the user.
-
-    Sourced from top-level frontmatter keys (Option B: flat dispatch
-    shape), not from a nested ``schedule:`` block.
-    """
-
-    interval: str = "manual"
-    schedule_hour: int = 9
-    schedule_day: int = 1
-    mode: str | None = None
-    model: str | None = None
-    reasoning_effort: str | None = None
-
-
-def _parse_workflow_frontmatter(meta: dict) -> Schedule:
-    """Validate top-level workflow frontmatter keys and return a ``Schedule``.
-
-    Reads ``interval``, ``schedule_hour``, ``schedule_day``, ``mode``,
-    ``model``, ``reasoning_effort`` directly from the prompt's
-    frontmatter.  ``interval`` defaults to ``"manual"`` when any other
-    execution-shape key is present but ``interval`` is omitted -- a
-    manual-only workflow is the conservative default given the Copilot
-    App's universal "run now" affordance.
-
-    Raises ``ValueError`` (with a human-readable message) on any
-    out-of-range or wrong-type field.  ``mode: autopilot`` is rejected
-    here with a targeted diagnostic before it can hit the DB layer's
-    generic CHECK violation -- third-party packages cannot ship
-    autopilot prompts; the user must opt in from the App UI.
-    """
-    if not isinstance(meta, dict):
-        raise ValueError("frontmatter must be a mapping")
-
-    interval = str(meta.get("interval", "manual"))
-    if interval not in _VALID_SCHEDULE_INTERVALS:
-        raise ValueError(
-            f"interval must be one of {sorted(_VALID_SCHEDULE_INTERVALS)}, got {interval!r}"
-        )
-
-    hour = meta.get("schedule_hour", 9)
-    if not isinstance(hour, int) or not (0 <= hour <= 23):
-        raise ValueError(f"schedule_hour must be int 0..23, got {hour!r}")
-
-    day = meta.get("schedule_day", 1)
-    if not isinstance(day, int) or not (0 <= day <= 6):
-        raise ValueError(f"schedule_day must be int 0..6, got {day!r}")
-
-    mode = meta.get("mode")
-    if mode is not None:
-        mode = str(mode)
-        if mode == "autopilot":
-            raise ValueError(
-                "mode 'autopilot' is not accepted via apm install -- "
-                "APM does not deploy workflows on autopilot. "
-                "Set autopilot manually in the Copilot App after enabling the row."
-            )
-        if mode not in _VALID_SCHEDULE_MODES:
-            raise ValueError(f"mode must be one of {sorted(_VALID_SCHEDULE_MODES)}, got {mode!r}")
-
-    model = meta.get("model")
-    if model is not None and not isinstance(model, str):
-        raise ValueError(f"model must be a string, got {model!r}")
-
-    reasoning_effort = meta.get("reasoning_effort")
-    if reasoning_effort is not None and not isinstance(reasoning_effort, str):
-        raise ValueError(f"reasoning_effort must be a string, got {reasoning_effort!r}")
-
-    return Schedule(
-        interval=interval,
-        schedule_hour=hour,
-        schedule_day=day,
-        mode=mode,
-        model=model,
-        reasoning_effort=reasoning_effort,
-    )
-
-
-# Back-compat alias retained for test imports; new code should use
-# ``_parse_workflow_frontmatter`` directly.
-_parse_schedule = _parse_workflow_frontmatter
-
-
-def _derive_package_owner(package_info) -> str:
-    """Best-effort owner-segment extraction for namespacing workflow ids.
-
-    Looks at the package's ``source`` (GitHub-style ``owner/repo`` or
-    URL) first, then ``author``, then falls back to ``"local"`` for
-    locally-sourced packages.  The returned string is slugified by the
-    DB-side ``namespaced_id`` helper, so any input is safe.
-    """
-    pkg = package_info.package
-    source = getattr(pkg, "source", None)
-    if isinstance(source, str) and source:
-        # github:foo/bar, https://github.com/foo/bar, foo/bar
-        s = source.split("://", 1)[-1]
-        s = s.split(":", 1)[-1]
-        parts = [p for p in s.split("/") if p and p != "github.com"]
-        if parts:
-            return parts[0]
-    author = getattr(pkg, "author", None)
-    if isinstance(author, str) and author.strip():
-        return author.strip()
-    return "local"
+__all__ = [
+    "PromptIntegrator",
+    "Schedule",
+    "_derive_package_owner",
+    "_is_workflow_shape",
+    "_parse_schedule",
+    "_parse_workflow_frontmatter",
+]
